Change default return type of Solve

+++ visual-studio-csharp/VainZero.Procon.CSharp/Program.cs
@@ -193,7 +193,7 @@ public partial class Program
 
 public sealed partial class Program
 {
-    int Solve()
+    long Solve()
     {
         return 0;
     }
